requirements.txt
asgiref==3.3.1
certifi==2020.6.20
cycler==0.10.0
decorator==4.4.2
Django==3.1.3
imageio==2.9.0
imutils==0.5.3
kiwisolver==1.3.1
matplotlib==3.3.2
networkx==2.5
numpy==1.19.4
opencv-contrib-python==4.2.0.34
pandas==1.1.4
pdf2image==1.14.0
Pillow==8.0.1
psycopg2==2.8.6
pyparsing==2.4.7
pytesseract==0.3.6
python-dateutil==2.8.1
pytz==2020.4
PyWavelets==1.1.1
scikit-image==0.17.2
scipy==1.5.3
six==1.15.0
sqlparse==0.4.1
tifffile==2020.10.1
